Added all custom changes from testing vault
I had been working on this project inside a separate testing folder.
This commit is an import of all of the code I wrote from that folder.
The only two difference is that I added 'CUSTOM' to all  places where I forgot,
now all the pieces of code I changed has that so it can be identified,
and the second change is that I removed some stuff that I no longer need.